update toggle favorite

+++ src/app/movies/details/details.component.ts
@@ -74,20 +74,25 @@ export class MovieDetailsComponent {
   }
 
   async toggleFavorite(showId: number, mediaType: string): Promise<void> {
-    this.snackbar.showLoading(true)
-    try {
-      if (this.isFavorite(showId, mediaType)) {
-        const favourite = await this.movieService.removeFavourite(this.userInfo.uuid, showId, mediaType);
-        this.usersData = this.globalService.getUsersData();
-      } else {
-        const favourite = await this.movieService.addFavourite(this.userInfo.uuid, showId, mediaType);
-        this.usersData = this.globalService.getUsersData();
+    if (this.globalService.getToken()) {
+      this.snackbar.showLoading(true)
+      try {
+        if (this.isFavorite(showId, mediaType)) {
+          const favourite = await this.movieService.removeFavourite(this.userInfo.uuid, showId, mediaType);
+          this.usersData = this.globalService.getUsersData()
+        } else {
+          const favourite = await this.movieService.addFavourite(this.userInfo.uuid, showId, mediaType);
+          this.usersData = this.globalService.getUsersData()
+        }
+      } catch (error) {
+        console.error('Error toggling favorite:', error);
+        this.snackbar.show('Error toggling favorite');
       }
-    } catch (error) {
-      console.error('Error toggling favorite:', error);
-      this.snackbar.show('Error toggling favorite');
-    } finally {
-      this.snackbar.showLoading(false)
+      finally {
+        this.snackbar.showLoading(false)
+      }
+    } else {
+      this.snackbar.show('Need to login first');
     }
   }
